+++ src/algorithm/math/divingBoard/divingBoard.ts
@@ -1,19 +1,10 @@
 export function divingBoard(shorter: number, longer: number, count: number): number[] {
-  [
-    shorter,
-    longer,
-  ] = [
-    shorter,
-    longer,
-  ].sort();
-
   if (count <= 0) {
     return [];
   }
 
   if (shorter === longer) {
-    return Array(count)
-      .fill(shorter);
+    return [shorter * count];
   }
 
   const result = [];