+++ src/algorithm/math/missingNumber/__test__/findErrorNums.test.ts
@@ -2,6 +2,19 @@ import { findErrorNums } from '../findErrorNums';
 
 test('findErrorNums', () => {
   expect(findErrorNums([
+    3,
+    2,
+    3,
+    4,
+    6,
+    5,
+  ]))
+    .toEqual([
+      3,
+      1,
+    ]);
+
+  expect(findErrorNums([
     1,
     2,
     4,