+++ src/algorithm/math/bits/countSetBits.ts
@@ -3,8 +3,8 @@ export function countSetBits(x: number): number {
   let number = x;
 
   while (number) {
-    setBitsCount += (number & 1);
-    number >>= 1;
+    number &= (number - 1);
+    setBitsCount++;
   }
 
   return setBitsCount;