+++ CHANGELOG.md
@@ -195,7 +195,7 @@
 - **math:** add pascalTriangle complexNumber fastPowering integerPartition liuHui ([cca718d](https://github.com/air-supply94/js-algorithm/commit/cca718d3ff1c0d85b4e1e64315b3a5571a909938))
 - **math:** add primalityTest and sieve ([9f3a1b7](https://github.com/air-supply94/js-algorithm/commit/9f3a1b72ecbd60857b3a71298bc34c7c831590f1))
 - **math:** update divingBoard ([d990253](https://github.com/air-supply94/js-algorithm/commit/d990253df88619a5a1091abccd8e64311cc3311e))
-- **math:** update fastPowering ([33cf620](https://github.com/air-supply94/js-algorithm/commit/33cf6203d5ef78a5235c2980474b44f7ef286ec7))
+- **math:** update fastPowering ([c193d7a](https://github.com/air-supply94/js-algorithm/commit/c193d7a343827c35ab52fbe25b19882b849c3393))
 - **math/bits:** add ([8bde116](https://github.com/air-supply94/js-algorithm/commit/8bde116141bc3a86625a2a3a44cdaf79ec6f535e))
 - **math/bits:** test ([3de3260](https://github.com/air-supply94/js-algorithm/commit/3de32608ced958befed76d42b1e8052c2fa6f6cc))
 - **mergeSort:** 更新 mergeSortedArrays concat 方式 ([8baf586](https://github.com/air-supply94/js-algorithm/commit/8baf5861287c1e603d619609628544db119c6781))