+++ CHANGELOG.md
@@ -123,6 +123,7 @@
 - **double-linked-list:** 增加排序 ([94af395](https://github.com/air-supply94/js-algorithm/commit/94af395a7a0085f805427b45864883f42e247350))
 - **doubleLinkedList:** add appendNode prependNode ([d88107b](https://github.com/air-supply94/js-algorithm/commit/d88107b33bed0ef76c1dc9baf4d5aed0c7a3e639))
 - **doubleLinkedList:** add deleteNode ([bba2af5](https://github.com/air-supply94/js-algorithm/commit/bba2af5ca0180e4ac2ee14ade28da7da035f76c7))
+- **doubleLinkedList:** add simple(刷题用) ([9076738](https://github.com/air-supply94/js-algorithm/commit/9076738156a5e54b252e11521508de4b26fdcd3e))
 - **doubleLinkedList:** remove deleteDumplicates swapPairs partion ([3fc6d5a](https://github.com/air-supply94/js-algorithm/commit/3fc6d5ad521bab55be4bf9e4f06d6a03c1bf9391))
 - **doubleLinkedList:** update reverse ([5a3b003](https://github.com/air-supply94/js-algorithm/commit/5a3b003fa8567ca478a4cf456867212533045574))
 - **doubleLinkedList:** update reverse ([0f3f7ef](https://github.com/air-supply94/js-algorithm/commit/0f3f7ef605953d15c7b844e478a9795d20802e27))